requirements.txt
apache-airflow==2.5.3
pyspark==3.4.1
psycopg2-binary==2.9.6
pendulum==2.1.2
Flask-Session==0.4.0
SQLAlchemy<2.0
connexion==2.14.1
praw
boto3
pysftp
pandas
pytest
great_expectations
streamlit==1.22.0
networkx==2.8.8
matplotlib==3.7.1
jupyter==1.0.0
nbformat==5.7.0
nbconvert==7.3.1